requirements.txt
aiofiles==0.7.0
click==8.0.1
Flask==2.0.1
httptools==0.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
orjson==3.6.3
sanic==21.6.2
sanic-routing==0.7.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
synonyms==3.16.0
threadpoolctl==2.2.0
ujson==4.1.0
uvloop==0.16.0
websockets==9.1
Werkzeug==2.0.1

PyYAML~=5.4.1
jieba~=0.42.1